apps/backend/requirements.txt
# Core frameworks
fastapi
uvicorn

# HTTP and async communication
httpx

# Firebase Admin SDK
firebase-admin

# Request parsing and validation
pydantic

# Environment config (optional, but useful)
python-dotenv

# RunPod serverless SDK (only needed in serverless mode)
runpod